Implemented the send_information_frame() and receive_information_frame() functions. This shit needs a HEAVY refactoring.

+++ lab2/link_layer.h
@@ -9,7 +9,7 @@
 typedef struct link_layer_t
 {
 	unsigned char *frame;
-	unsigned long frameSize;
+	int frameSize;
 	unsigned char sequenceNumber;
 } LinkLayer;
